Refactor code into packages and add enums to Commands.java
