+++ src/main/java/sentinel/storage/Storage.java
@@ -1,3 +1,9 @@
+package sentinel.storage;
+
+import sentinel.parser.Parser;
+import sentinel.task.Task;
+import sentinel.task.TaskList;
+
 import java.io.File;
 import java.nio.file.Files;
 import java.io.FileWriter;
@@ -7,7 +13,13 @@ import java.util.Scanner;
 
 import java.io.IOException;
 
+/**
+ * Represents a storage for storing Sentinel's tasks.
+ */
 public class Storage {
+        /**
+         * Loads the task list that was saved.
+         */
         public static TaskList load() throws IOException {
             File f = null;
 
@@ -42,6 +54,9 @@ public class Storage {
             return listOfTasks;
         }
 
+        /**
+         * Saves the current task list to a file.
+         */
         public static void save(String content) throws IOException {
             FileWriter fw = new FileWriter("src/main/data/data.txt");
             fw.write(content);
