+++ src/main/java/sentinel/task/TaskList.java
@@ -1,3 +1,7 @@
+package sentinel.task;
+
+import sentinel.SentinelException;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 
@@ -24,7 +28,7 @@ public class TaskList {
      * Adds a todo to the list.
      *
      * @param description Description of the todo.
-     * @return Todo that was added.
+     * @return sentinel.task.Todo that was added.
      */
     public Task addTodo(String description) {
         Task newTodo = new Todo(description);
@@ -39,7 +43,7 @@ public class TaskList {
      * @param description Description of the event.
      * @param startTime Starting time for the event timeframe.
      * @param endTime Ending time for the event timeframe.
-     * @return Event that was added.
+     * @return sentinel.task.Event that was added.
      */
     public Task addEvent(String description, LocalDate startTime, LocalDate endTime) {
         Task newEvent = new Event(description, startTime, endTime);
@@ -53,7 +57,7 @@ public class TaskList {
      *
      * @param description Description of the deadline.
      * @param endTime Time to do the deadline by.
-     * @return Deadline that was added.
+     * @return sentinel.task.Deadline that was added.
      */
     public Task addDeadline(String description, LocalDate endTime) {
         Task newDeadline = new Deadline(description, endTime);
@@ -66,7 +70,7 @@ public class TaskList {
      * Marks the task on the list as done.
      *
      * @param taskNumber Number of the task.
-     * @return Task that was done.
+     * @return sentinel.task.Task that was done.
      * @throws SentinelException if the task does not exist.
      */
     public Task markAsDone(int taskNumber) throws SentinelException {
@@ -83,7 +87,7 @@ public class TaskList {
      * Marks the task on the list as undone.
      *
      * @param taskNumber Number of the task.
-     * @return Task that was undone.
+     * @return sentinel.task.Task that was undone.
      * @throws SentinelException if the task does not exist.
      */
     public Task markAsUndone(int taskNumber) throws SentinelException {
@@ -109,7 +113,7 @@ public class TaskList {
      * Deletes the task on the list.
      *
      * @param taskNumber Number of the task.
-     * @return Task that was deleted.
+     * @return sentinel.task.Task that was deleted.
      * @throws SentinelException if the task does not exist.
      */
     public Task deleteTask(int taskNumber) throws SentinelException {
