+++ src/main/java/sentinel/task/Todo.java
@@ -1,3 +1,5 @@
+package sentinel.task;
+
 /**
  * Represents a todo task.
  */
